requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.7.0
aioredis==1.3.1
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
async-timeout==3.0.1
attrs==25.1.0
autopep8==2.3.2
calmsize==0.1.3
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.0.4
clip==0.2.0
colorlog==6.9.0
comm==0.2.2
configargparse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
docstring-parser==0.16
einops==0.3.0
embreex==2.17.7.post6
eval-type-backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
flask==2.2.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
# fused-ssim==0.0.0
grpcio==1.70.0
gsplat==1.4.0
h11==0.14.0
h5py==3.13.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jaxtyping==0.2.36
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
kornia==0.4.1
lazy-loader==0.4
lightning-utilities==0.12.0
loguru==0.7.3
lpips==0.1.4
lxml==5.3.1
manifold3d==3.0.1
mapbox-earcut==1.0.3
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
# mobile-sam==1.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
narwhals==1.29.0
natsort==8.4.0
nbformat==5.10.4
nerfview==0.0.3
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickle5==0.0.11
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plyfile==1.1
prompt-toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycodestyle==2.12.1
pycollada==0.9
pycolmap==0.2.0
pydegensac==0.1.2
pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
# pytorch-lightning==1.9.0
pytorch-memlab==0.3.0
pytz==2025.1
pyyaml==6.0.2
# ray==1.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
rtree==1.3.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
supervision==0.25.1
svg-path==6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorly==0.9.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.0
tomli==2.2.1
# torch==1.13.1+cu117
# torch-fidelity==0.3.0
# torchaudio==0.13.1+cu117
# torchmetrics==1.0.0
# torchvision==0.14.1+cu117
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
transforms3d==0.4.2
trimesh==4.6.4
triton==3.2.0
typeguard==4.4.2
typing-extensions==4.12.2
tyro==0.9.16
tzdata==2025.1
ultralytics==8.0.120
ultralytics-thop==2.0.14
urllib3==2.3.0
vhacdx==0.0.8.post2
virtualenv==20.29.2
viser==0.2.23
wcwidth==0.2.13
websockets==15.0
werkzeug==3.0.6
widgetsnbextension==4.0.13
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
yourdfpy==0.0.57
zipp==3.21.0